+++ IsekaiMod/Content/Features/ExceptionalFeats/ExceptionalFeatSelection.cs
@@ -8,9 +8,11 @@ namespace IsekaiMod.Content.Features.ExceptionalFeats
 {
     class ExceptionalFeatSelection
     {
+        private static readonly BlueprintFeatureSelection MythicFeatSelection = Resources.GetBlueprint<BlueprintFeatureSelection>("9ee0f6745f555484299b0a1563b99d81");
+        private static readonly BlueprintFeatureSelection ExtraFeatMythicFeat = Resources.GetBlueprint<BlueprintFeatureSelection>("e10c4f18a6c8b4342afe6954bde0587b");
+        private static readonly BlueprintFeatureSelection ExtraMythicAbilityMythicFeat = Resources.GetBlueprint<BlueprintFeatureSelection>("8a6a511c55e67d04db328cc49aaad2b8");
         public static void Add()
         {
-            var BasicFeatSelection = Resources.GetBlueprint<BlueprintFeatureSelection>("247a4068296e8be42890143f451b4b45");
             var ExceptionalFeatSelection = Helpers.CreateBlueprint<BlueprintFeatureSelection>("ExceptionalFeatSelection", bp => {
                 bp.SetName("Exceptional Feats");
                 bp.SetDescription("Exceptional feats are feats that no ordinary NPC possess.\nSource: Isekai Mod");
@@ -20,11 +22,14 @@ namespace IsekaiMod.Content.Features.ExceptionalFeats
                 bp.Group = FeatureGroup.Feat;
                 bp.Group2 = FeatureGroup.TricksterFeat;
 
-                // Add Exceptional feats later
-                bp.m_AllFeatures = new BlueprintFeatureReference[0];
+                // Add more Exceptional feats later
+                bp.m_AllFeatures = MythicFeatSelection.m_AllFeatures
+                    .RemoveFromArray(ExtraFeatMythicFeat.ToReference<BlueprintFeatureReference>())
+                    .RemoveFromArray(ExtraMythicAbilityMythicFeat.ToReference<BlueprintFeatureReference>());
             });
 
             // Add to Basic Feat Selection
+            var BasicFeatSelection = Resources.GetBlueprint<BlueprintFeatureSelection>("247a4068296e8be42890143f451b4b45");
             BasicFeatSelection.m_AllFeatures = BasicFeatSelection.m_AllFeatures.AddToFirst(ExceptionalFeatSelection.ToReference<BlueprintFeatureReference>());
         }
         public static void AddToSelection(BlueprintFeature feature)